// src/types/ProductType.ts
export interface ProductType {
    brand_name: string
    description: string
    imt_name: string
    nm_colors_names: string
    nm_id: 118982
    subj_name: string
    subj_root_name: string
    vendor_code: "82992/beige"
    backgroundColor?: string; 
}

export interface ProductsTypeObject {
    row: ProductType
}